escalating missing library to FATAL_ERROR
It was previously a WARNING in CMake, but that
leads to missing symbol errors, which can be
misleading since the library was actually not
found but the first inclination is to check the
library which contains the symbols for errors.

We might consider the need to change this back
in the future for cases where having the library
is not critical.