Always write output as utf-8.
CMake documentation suggests that we should be writing 7-bit ascii
CMake source files or writing UTF-8 with a byte order mark. (Source:
https://cmake.org/cmake/help/v3.5/manual/cmake-language.7.html#encoding).

This doesn't actually do either of those things. It just cements our
position of non-compliance (writing utf-8 without a byte order mark)
so that builds don't crash if the system encoding is other than utf-8.

Alternatively we could sanitize the generated CMake content so it is
7-bit ascii and explicitly write it as such or consider adding the byte
order mark.