Report skipped tests in CTest output (#243)
When adding a test using `ament_add_test`, the `SKIP_TEST` argument
results in the `--skip-test` argument being passed to the test wrapper
script `run_test.py`. The wrapper script then writes a JUnit output
describing that the test was skipped, and returns 0.

As far as CTest knows, the test succeeded and shows `Passed` on the
console. However, since we know that the test will be skipped by the
wrapper, and we expect the wrapper to return 0 after it writes the JUnit
file, we can set a test property that will mark the test as `Skipped`
when the wrapper returns 0.

This way, the JUnit output file is still written, but CTest displays the
test as skipped as well.

Signed-off-by: Scott K Logan <logans@cottsay.net>
Signed-off-by: Jose Luis Rivero <jrivero@osrfoundation.org>