Add SKIP_RETURN_CODE argument to ament_add_test
This makes the `run_test.py` wrapper aware of the `SKIP_RETURN_CODE`
property on CTest tests. In the existing implementation, the wrapper
detects that no result file was generated and overrides the special
return code coming from the test, making the the CTest feature fail
completely.

This change makes the wrapper script aware of the special return code,
and when detected, will write a 'skipped' result file instead of a
'failed' result file, and pass along the special return code as-is. Now
the gtest result and the ctest results both show the test as 'skipped'
when the special return flag is used.

Note that none of this behavior is enabled by default, which is
important because we wouldn't want a test to fail and return a code
which we've decided is the special 'skip' return code. Only tests which
are aware of this feature should use it.

Signed-off-by: Scott K Logan <logans@cottsay.net>