Skip performance tests using a CMake variable
These tests can be fairly heavy, so we don't want to run them by
default. It would be better if there was a way to skip the tests by
default in such a way that they could be specifically un-skipped at
runtime, but I can't find a mechanism in CMake or CTest that would allow
us to achieve that behavior without leveraging environment variables.

Signed-off-by: Scott K Logan <logans@cottsay.net>