Introduce relocatable AMENT_IMPORT_PREFIX
This moves the logic to compute the install prefix to a single location
in the main config file. Extra config files can use the new
AMENT_IMPORT_PREFIX variable, which simplifies the code and makes it
more robust to future changes.

The logic to compute AMENT_IMPORT_PREFIX has been taken from the
CMakePackageConfigHelpers module. The code is well-tested and also
handles symlink aliasing.

Signed-off-by: Timo Röhling <roehling@debian.org>