ament_python_install_package: exclude libraries from install
hen rosidl_python is run to generate messages, the
lib<msgname>__rosidl_generator_py.so (in case of linux) is installed
into PYTHON_INSTALL_DIR and lib. This is a regression from foxy to humble.
In the past, only .py files were copied, now we copy over the complete
directory, which happens to also have the above library in it.

Since I do not know better, I exclude libraries from being copied.

Signed-off-by: Matthias Schoepfer <m.schoepfer@rethinkrobotics.com>