Use file(GENERATE OUTPUT) to create dsv files
Using file(WRITE) and file(APPEND) causes the modification stamp of the
file to be changed each time CMake configures, resluting in an
'Installing' message rather than an 'Up-to-date' message even though the
file content is identical.

Using file(GENERATE OUTPUT) updates the timestamp of the file only if
the content changes.

Signed-off-by: Scott K Logan <logans@cottsay.net>