Specify scope of link libraries
We know these libraries are only build requirements and not usage
requirements so we can explicitly specify as such. This will allow
users of these targets to also use these scope keywords in their
own target_link_libraries calls instead of being forced to omit them
as is the case currently.

Signed-off-by: Chris Thrasher <chrisjthrasher@gmail.com>