Fix CMake --find-package mode
There is a CMake mode which can be used to extract basic compile and
linker flags as would be produced by a call to find_package(). Existing
ament_cmake code always assumes that we're being invoked under a
"PROJECT" role. This change skips certain project-specific parts of
ament_cmake_core and ament_cmake_test which shouldn't be run unless
we're actually configuring a CMakeLists.txt project.

The smoke tests added by this change validate that the "extra" CMake
functionality provided by these packages continues to be loadable
outside of a project context. It does not validate that the defined
functions can actually be used, but will serve as a validation that
features like CMake --find-package should work.

Signed-off-by: Scott K Logan <logans@cottsay.net>